# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Broker API",
    "version": "1.0.0",
    "description": "This is the APIs documentation for Broker Platform"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" },
    { "url": "https://broker-six-navy.vercel.app/", "description": "Production server" }
  ],
...
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Sign Up",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent to email for verification"
          },
          "400": {
            "description": "Missing or invalid fields"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify OTP for user email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "csdNumber": {
                      "type": "string",
                      "description": "Generated CSD number in the format YYYYMMDD {gender code} {country code}{serial}",
                      "example": "20251105 1 RW01"
                    }
                  },
                  "required": ["message", "csdNumber"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP or email"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/company": {
      "get": {
        "tags": ["Company"],
        "summary": "List companies",
        "description": "Returns companies created on the platform. Public endpoint; supports simple text filtering.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Optional query to filter by name, ticker, or sector"
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompanyListResponse" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "tags": ["Company"],
        "summary": "Create a new company",
        "description": "Requires SUPER_ADMIN or COMPANY role. Company users become owners of the companies they create.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompanyCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "400": { "description": "Invalid company data" },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "409": { "description": "Conflict - ticker already exists" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/company/{id}": {
      "get": {
        "tags": ["Company"],
        "summary": "Get company by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Company details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "404": { "description": "Company not found" },
          "500": { "description": "Server error" }
        }
      },
      "patch": {
        "tags": ["Company"],
        "summary": "Update company by ID",
        "description": "Requires SUPER_ADMIN or COMPANY role. Company users can only modify companies they created.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompanyUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "400": { "description": "Invalid company data" },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role or ownership" },
          "404": { "description": "Company not found" },
          "409": { "description": "Conflict - ticker already exists" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "tags": ["Company"],
        "summary": "Delete company by ID",
        "description": "Requires SUPER_ADMIN or COMPANY role. Company users can only remove companies they created.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Company deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role or ownership" },
          "404": { "description": "Company not found" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "minLength": 3 },
          "lastName": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "phoneCountryCode": { "type": "string", "pattern": "^\\+[0-9]{1,4}$" },
          "phone": {
            "type": "string",
            "pattern": "^[0-9]{4,15}$",
            "description": "Digits only; spaces are removed before validation."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Must contain uppercase, lowercase, numeric, and special characters."
          },
          "confirmPassword": { "type": "string", "minLength": 8 },
          "idNumber": { "type": "string", "minLength": 1 },
          "csdNumber": { "type": "string", "description": "Optional Central Securities Depository number" },
          "passportPhoto": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's passport photo"
          },
          "idDocument": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's identification document"
          },
          "dateOfBirth": { "type": "string", "format": "date" },
          "gender": {
            "type": "string",
            "enum": ["male", "female"]
          },
          "country": { "type": "string", "minLength": 1 },
          "city": { "type": "string", "minLength": 1 },
          "occupation": { "type": "string", "minLength": 1 },
          "investmentExperience": { "type": "string", "minLength": 1 }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phoneCountryCode",
          "phone",
          "password",
          "confirmPassword",
          "idNumber",
          "passportPhoto",
          "idDocument",
          "dateOfBirth",
          "gender",
          "country",
          "city",
          "occupation",
          "investmentExperience"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": ["SUPER_ADMIN", "ADMIN", "AGENT", "COMPANY", "CLIENT"]
      },
...
      "CompanyOwnerSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/UserRole" }
        },
        "required": ["id", "firstName", "lastName", "email", "role"]
      },
      "Company": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "ticker": {
            "type": "string",
            "nullable": true,
            "description": "Uppercase ticker symbol"
          },
          "description": { "type": "string", "nullable": true },
          "sector": { "type": "string", "nullable": true },
          "sharePrice": {
            "type": "string",
            "nullable": true,
            "description": "String representation of the company's share price"
          },
          "totalShares": { "type": "integer", "nullable": true },
          "availableShares": {
            "type": "integer",
            "nullable": true,
            "description": "Available float of shares. Must not exceed totalShares when both are present."
          },
          "createdBy": {
            "$ref": "#/components/schemas/CompanyOwnerSummary",
            "nullable": true
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "closingPrice": {
            "type": "string",
            "nullable": true,
            "description": "End-of-day closing price from the latest market snapshot."
          },
          "previousClosingPrice": {
            "type": "string",
            "nullable": true,
            "description": "Previous session closing price."
          },
          "priceChange": {
            "type": "string",
            "nullable": true,
            "description": "Formatted price change (e.g. +0.65%)."
          },
          "tradedVolume": {
            "type": "string",
            "nullable": true,
            "description": "Latest session traded volume. Represented as string to support large values."
          },
          "tradedValue": {
            "type": "string",
            "nullable": true,
            "description": "Latest session traded value. Represented as string to preserve precision."
          },
          "snapshotDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Market snapshot reference timestamp."
          },
          "contract": {
            "type": "string",
            "nullable": true,
            "description": "Optional contract or prospectus reference URL."
          },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "createdAt", "updatedAt"]
      },
      "CompanyListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Company" }
          }
        },
        "required": ["data"]
      },
      "CompanyResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/Company" }
        },
        "required": ["data"]
      },
      "CompanyCreateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "ticker": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10,
            "description": "Unique uppercase ticker symbol"
          },
          "description": { "type": "string", "nullable": true },
          "sector": { "type": "string", "nullable": true },
          "sharePrice": {
            "type": "string",
            "nullable": true,
            "description": "Optional decimal share price. Provide as string to preserve precision."
          },
          "totalShares": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "availableShares": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "description": "Must not exceed totalShares when both are provided"
          },
          "closingPrice": {
            "type": "string",
            "nullable": true,
            "description": "End-of-day closing price from the latest market snapshot."
          },
          "previousClosingPrice": {
            "type": "string",
            "nullable": true,
            "description": "Previous session closing price."
          },
          "priceChange": {
            "type": "string",
            "nullable": true,
            "description": "Formatted price change (e.g. +0.65%)."
          },
          "tradedVolume": {
            "type": "string",
            "nullable": true,
            "description": "Latest session traded volume. Represented as string to support large values."
          },
          "tradedValue": {
            "type": "string",
            "nullable": true,
            "description": "Latest session traded value. Represented as string to preserve precision."
          },
          "snapshotDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Market snapshot reference timestamp."
          },
          "contract": {
            "type": "string",
            "nullable": true,
            "description": "Optional contract or prospectus reference URL."
          }
        },
        "required": ["name", "ticker"]
      },
      "CompanyUpdateRequest": {
        "type": "object",
        "description": "At least one field must be provided",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "ticker": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          },
          "description": { "type": "string", "nullable": true },
          "sector": { "type": "string", "nullable": true },
          "sharePrice": {
            "type": "string",
            "nullable": true,
            "description": "Optional decimal share price. Provide as string to preserve precision."
          },
          "totalShares": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "availableShares": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "description": "Must not exceed totalShares when both are provided"
          },
          "closingPrice": {
            "type": "string",
            "nullable": true
          },
          "previousClosingPrice": {
            "type": "string",
            "nullable": true
          },
          "priceChange": {
            "type": "string",
            "nullable": true
          },
          "tradedVolume": {
            "type": "string",
            "nullable": true
          },
          "tradedValue": {
            "type": "string",
            "nullable": true
          },
          "snapshotDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "contract": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        },
        "required": ["message"]
      },
...
      "VerifyOtpRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "otp": { "type": "string", "pattern": "^\\d{6}$" }
        },
        "required": ["email", "otp"]
      },
...
    }
  }
}